Revert "Merge commit '319f3a1ecb19b5dcfbe4a2b665a9a600b96243fe' into fleet-composition"
This reverts commit e2a74672093cd18924b3dbfa9cfb2fab82e95642, reversing
changes made to 7814a9e7d4958c0c624bff24c5a109c8f72dcb7e.

+++ OgameGenSim/Classes/CharacterClassBooster.cs
@@ -5,9 +5,9 @@ namespace OgameGenSim.Classes;
 public class CharacterClassBooster
 {
     [JsonPropertyName("1")]
-    public double Collector { get; set; }
+    public int Collector { get; set; }
     [JsonPropertyName("2")]
-    public double General { get; set; }
+    public int General { get; set; }
     [JsonPropertyName("3")]
-    public double Discoverer { get; set; }
+    public int Discoverer { get; set; }
 }
